+++ src/game/tests.rs
@@ -360,6 +360,7 @@ mod vintage_cube_chainsaw;
 mod vintage_cube_chandra;
 mod vintage_cube_chromatic_star;
 mod vintage_cube_chrome_mox;
+mod vintage_cube_city_of_traitors;
 mod vintage_cube_collective_brutality;
 mod vintage_cube_colonnade;
 mod vintage_cube_concealing_curtains;